Changed datatype to DOUBLE

+++ src/migrations/20210817151308-cashback.js
@@ -10,11 +10,11 @@ module.exports = {
       },
       value: {
         allowNull: false,
-        type: Sequelize.DECIMAL,
+        type: Sequelize.DOUBLE,
       },
       percentage: {
         allowNull: false,
-        type: Sequelize.FLOAT,
+        type: Sequelize.DOUBLE,
       },
       purchase_uid: {
         allowNull: false,
